+++ mashid/__init__.py
@@ -1,6 +1,6 @@
 """mashID: identify organisms from genome assemblies or raw reads using Mash."""
 
-__version__ = "0.2.5"
+__version__ = "0.2.6"
 __author__ = "Marc-Olivier Duceppe"